+++ modules-core/prov-model/src/main/java/org/openprovenance/prov/bookptm/DerivationBuilder.java
@@ -7,7 +7,7 @@ import org.openprovenance.prov.model.QualifiedName;
 import org.openprovenance.prov.model.builder.Builder;
 import org.openprovenance.prov.model.builder.Prefix;
 
-import java.io.IOException;
+import static org.openprovenance.prov.model.NamespacePrefixMapper.PROV_EXT_NS;
 
 public class DerivationBuilder {
 
@@ -21,6 +21,14 @@ public class DerivationBuilder {
         Builder builder = new Builder(pFactory, pFactory, pFactory);
         Definitions defs = getDefinitions(builder);
         transportingDescription(builder, defs);
+        Document doc=builder.build();
+        return doc;
+    }
+
+    public Document makeDocument_TransportingFullTriangle() {
+        Builder builder = new Builder(pFactory, pFactory, pFactory);
+        Definitions defs = getDefinitions(builder);
+        transportingDescriptionFullTriangle(builder, defs);
         Document doc=builder.build();
         return doc;
     }
@@ -44,7 +52,15 @@ public class DerivationBuilder {
     public Document makeDocument_Attribution() {
         Builder builder = new Builder(pFactory, pFactory, pFactory);
         Definitions defs = getDefinitions(builder);
-        attributionDescripton(builder, defs, true);
+        attributionDescription(builder, defs, true);
+        Document doc=builder.build();
+        return doc;
+    }
+
+    public Document makeDocument_AttributionFullTriangle() {
+        Builder builder = new Builder(pFactory, pFactory, pFactory);
+        Definitions defs = getDefinitions(builder);
+        attributionDescriptionFullTriangle(builder, defs, true);
         Document doc=builder.build();
         return doc;
     }
@@ -57,6 +73,15 @@ public class DerivationBuilder {
         return doc;
     }
 
+    public Document makeDocument_SpecializationFullTriangle() {
+        Builder builder = new Builder(pFactory, pFactory, pFactory);
+        Definitions defs = getDefinitions(builder);
+        specializationDescriptionFullTriangle(builder, defs, true);
+        Document doc=builder.build();
+        return doc;
+    }
+
+
     public Document makeDerivation() {
         Builder builder = new Builder(pFactory, pFactory, pFactory);
         return makeDerivation(builder);
@@ -68,7 +93,7 @@ public class DerivationBuilder {
 
         transportingDescription(builder, defs);
 
-        attributionDescripton(builder, defs, false);
+        attributionDescription(builder, defs, false);
 
         weighingDescription(builder, defs, false);
 
@@ -80,7 +105,7 @@ public class DerivationBuilder {
         return doc;
     }
 
-    private void attributionDescripton(Builder builder, Definitions defs, boolean standalone) {
+    private void attributionDescription(Builder builder, Definitions defs, boolean standalone) {
 
         if (standalone) {
 
@@ -99,6 +124,99 @@ public class DerivationBuilder {
         builder.wasAttributedTo()
                 .entity("b34/5")
                 .agent("luc")
+                .type(defs.Ownership)
+                .build();
+    }
+
+    private void attributionDescriptionFullTriangle(Builder builder, Definitions defs, boolean standalone) {
+
+        if (standalone) {
+
+            builder.entity()
+                    .id(defs.XID, "b34/5").aka()
+                    .type(defs.Box)
+                    .label("a box")
+                    .build();
+        }
+
+
+        builder.entity()
+                .id(defs.XID, "fpc045").aka()
+                .type(defs.FPC)
+                .label("flat-packed cardboard")
+                .build();
+
+        builder.entity()
+                .id(defs.XID, "bks643").aka()
+                //.type(defs.Books)
+                .label("books")
+                .build();
+
+
+        builder.agent()
+                .id(defs.XID, "luc").knownAsLocal()
+                .type(name.PROV_PERSON)
+                .build();
+
+        builder.activity()
+                .id(defs.XID, "a7667").knownAsLocal()
+                .label("packing activity")
+                .type(defs.Packing)
+                .build();
+
+        builder.wasGeneratedBy()
+                .id(defs.XID, "wgb0").aka()
+                .entity("b34/5")
+                .activity("a7667")
+                .build();
+
+        builder.used()
+                .id(defs.XID, "used1").aka()
+                .activity("a7667")
+                .entity("fpc045")
+                .build();
+
+        builder.used()
+                .id(defs.XID, "used2").aka()
+                .activity("a7667")
+                .entity("bks643")
+                .build();
+
+        builder.wasDerivedFrom()
+                .id(defs.XID, "deriv1").aka()
+                .generatedEntity("b34/5")
+                .usedEntity("fpc045")
+                .activity("a7667")
+                .generation("wgb0")
+                .usage("used1")
+                .type(defs.Packing)
+                .build();
+
+        builder.wasDerivedFrom()
+                .id(defs.XID, "deriv2").aka()
+                .generatedEntity("b34/5")
+                .usedEntity("bks643")
+                .activity("a7667")
+                .generation("wgb0")
+                .usage("used2")
+                .type(defs.Packing)
+                .build();
+
+
+        builder.wasAssociatedWith()
+                .id(defs.XID, "waw1").aka()
+                .activity("a7667")
+                .agent("luc")
+                .build();
+
+
+        builder.wasAttributedTo()
+                .entity("b34/5")
+                .agent("luc")
+                .type(defs.Ownership)
+                .attrQn(defs.association,"waw1")
+                .attrQn(defs.generation,"wgb0")
+                .attrQn(defs.activity, "a7667")
                 .build();
     }
 
@@ -119,14 +237,14 @@ public class DerivationBuilder {
                 .build();
 
         builder.entity()
-                .id(defs.XID, "b347").aka()
+                .id(defs.XID, "b34/7").aka()
                 .type(defs.Box)
                 .label("a weighted box")
                 .attr(defs.weight, "8kg")
                 .build();
 
         builder.wasGeneratedBy()
-                .entity("b347")
+                .entity("b34/7")
                 .activity("a7588")
                 .build();
 
@@ -136,7 +254,7 @@ public class DerivationBuilder {
                 .build();
 
         builder.wasDerivedFrom()
-                .generatedEntity("b347")
+                .generatedEntity("b34/7")
                 .usedEntity("b34/6")
                 .type(defs.Weighing)
                 .activity("a7588")
@@ -215,6 +333,60 @@ public class DerivationBuilder {
                 .build();
     }
 
+    private void transportingDescriptionFullTriangle(Builder builder, Definitions defs) {
+        builder.entity()
+                .id(defs.XID, "b34/5").aka()
+                .type(defs.Box)
+                .label("a box somewhere")
+                .location(defs.London)
+                .attr(defs.weight, "10kg")
+                .build();
+
+        builder.entity()
+                .id(defs.XID, "b34/6").aka()
+                .type(defs.Box)
+                .label("a box elsewhere")
+                .location(defs.Brighton)
+                .build();
+
+        builder.activity()
+                .id(defs.XID, "a7543").knownAsLocal()
+                .label("some activity")
+                .type(defs.Transporting)
+                .start("2024-07-16T10:05:00")
+                .end("2024-07-16T12:47:00")
+                .build();
+
+        builder.wasGeneratedBy()
+                .id(defs.XID, "wgb1").aka()
+                .entity("b34/6")
+                .activity("a7543")
+                .type(defs.DropOff)
+                .time("2024-07-16T11:58:00")
+                .label("drop-off at destination")
+                .build();
+
+        builder.used()
+                .id(defs.XID, "used1").aka()
+                .activity("a7543")
+                .entity("b34/5")
+                .type(defs.PickUp)
+                .time("2024-07-16T10:20:00")
+                .label("pick up at origin")
+                .build();
+
+        builder.wasDerivedFrom()
+                .generatedEntity("b34/6")
+                .usedEntity("b34/5")
+                .type(defs.Transporting)
+                .activity("a7543")
+                .generation("wgb1")
+                .usage("used1")
+                .build();
+
+    }
+
+
     public void transportingAttribution(Builder builder, Definitions defs, boolean standalone) {
         if (standalone) {
             builder.activity()
@@ -290,6 +462,52 @@ public class DerivationBuilder {
                 .specificEntity("b34/6")
                 .build();
     }
+    private void specializationDescriptionFullTriangle(Builder builder, Definitions defs, boolean standalone) {
+        if (standalone) {
+
+            builder.entity()
+                    .id(defs.XID, "b34/5").aka()
+                    .type(defs.Box)
+                    .label("a box somewhere")
+                    .location(defs.London)
+                    .attr(defs.weight, "10kg")
+                    .build();
+
+            builder.entity()
+                    .id(defs.XID, "b34/6").aka()
+                    .type(defs.Box)
+                    .label("a box elsewhere")
+                    .location(defs.Brighton)
+                    .build();
+
+            builder.wasDerivedFrom()
+                    .id(defs.XID,"deriv1").aka()
+                    .generatedEntity("b34/6")
+                    .usedEntity("b34/5")
+                    .type(defs.Transporting)
+                    .build();
+        }
+
+        builder.entity()
+                .id(defs.XID, "b34").aka()
+                .type(defs.Box)
+                .label("a box")
+                .build();
+
+        builder.specializationOf()
+                .id(defs.XID,"spe1").aka()
+                .generalEntity("b34")
+                .specificEntity("b34/5")
+                .build();
+
+        builder.specializationOf()
+                .generalEntity("b34")
+                .specificEntity("b34/6")
+                .attrQn(defs.specialization,"spe1")
+                .attrQn(defs.derivation,"deriv1")
+                .attrQn(defs.entity, "b34/5")
+                .build();
+    }
 
 
 
@@ -297,13 +515,18 @@ public class DerivationBuilder {
         Prefix VOCAB  = builder.prefix("vocab");
         Prefix XID    = builder.prefix("xid");
         Prefix FOAF    = builder.prefix("foaf");
+        Prefix PROVEXT = builder.prefix("provext");
         builder.prefix(VOCAB, EX_NS_URI);
         builder.prefix(XID, EX_IDS_URI);
         builder.prefix(FOAF, FOAF_URI);
+        builder.prefix(PROVEXT, PROV_EXT_NS);
 
         QualifiedName Box          = builder.qn(VOCAB, "Box");
+        QualifiedName FPC          = builder.qn(VOCAB, "FPC");
+        QualifiedName Ownership    = builder.qn(VOCAB, "Ownership");
         QualifiedName Transporting = builder.qn(VOCAB, "Transporting");
         QualifiedName Weighing     = builder.qn(VOCAB, "Weighing");
+        QualifiedName Packing     = builder.qn(VOCAB, "Packing");
         QualifiedName PickUp       = builder.qn(VOCAB, "PickUp");
         QualifiedName DropOff      = builder.qn(VOCAB, "DropOff");
         QualifiedName London       = builder.qn(VOCAB, "London");
@@ -313,17 +536,27 @@ public class DerivationBuilder {
         QualifiedName scientist    = builder.qn(VOCAB, "scientist");
         QualifiedName scale        = builder.qn(VOCAB, "scale");
         QualifiedName Instrument   = builder.qn(VOCAB, "Instrument");
+        QualifiedName specialization=builder.qn(PROVEXT, "specialization");
+        QualifiedName derivation    =builder.qn(PROVEXT, "derivation");
+        QualifiedName entity        =builder.qn(PROVEXT, "entity");
+        QualifiedName association   =builder.qn(PROVEXT, "association");
+        QualifiedName generation    =builder.qn(PROVEXT, "generation");
+        QualifiedName activity      =builder.qn(PROVEXT, "activity");
+
 
         QualifiedName foaf_name    = builder.qn(FOAF, "name");
-        Definitions result = new Definitions(XID, Box, Transporting, Weighing, PickUp, DropOff, London, Brighton, weight, foaf_name, driver, scientist, scale, Instrument);
+        Definitions result = new Definitions(XID, Box, FPC, Ownership, Transporting, Weighing, Packing, PickUp, DropOff, London, Brighton, weight, foaf_name, driver, scientist, scale, Instrument, specialization, derivation, entity, association, generation, activity);
         return result;
     }
 
     public static class Definitions {
         public final Prefix XID;
         public final QualifiedName Box;
+        public final QualifiedName FPC;
+        public final QualifiedName Ownership;
         public final QualifiedName Transporting;
         public final QualifiedName Weighing;
+        public final QualifiedName Packing;
         public final QualifiedName PickUp;
         public final QualifiedName DropOff;
         public final QualifiedName London;
@@ -334,12 +567,21 @@ public class DerivationBuilder {
         public final QualifiedName scientist;
         public final QualifiedName scale;
         public final QualifiedName Instrument;
+        public final QualifiedName specialization;
+        public final QualifiedName derivation;
+        public final QualifiedName entity;
+        public final QualifiedName association;
+        public final QualifiedName generation;
+        public final QualifiedName activity;
 
-        public Definitions(Prefix XID, QualifiedName Box, QualifiedName Transporting, QualifiedName Weighing, QualifiedName PickUp, QualifiedName DropOff, QualifiedName London, QualifiedName Brighton, QualifiedName weight, QualifiedName foaf_name, QualifiedName driver, QualifiedName scientist, QualifiedName scale, QualifiedName instrument) {
+        public Definitions(Prefix XID, QualifiedName Box, QualifiedName fpc, QualifiedName Ownership, QualifiedName Transporting, QualifiedName Weighing, QualifiedName packing, QualifiedName PickUp, QualifiedName DropOff, QualifiedName London, QualifiedName Brighton, QualifiedName weight, QualifiedName foaf_name, QualifiedName driver, QualifiedName scientist, QualifiedName scale, QualifiedName instrument, QualifiedName specialization, QualifiedName derivation, QualifiedName entity, QualifiedName association, QualifiedName generation, QualifiedName activity) {
             this.XID = XID;
             this.Box = Box;
+            this.FPC = fpc;
+            this.Ownership = Ownership;
             this.Transporting = Transporting;
             this.Weighing = Weighing;
+            this.Packing = packing;
             this.PickUp = PickUp;
             this.DropOff = DropOff;
             this.London = London;
@@ -350,6 +592,12 @@ public class DerivationBuilder {
             this.scientist = scientist;
             this.scale = scale;
             this.Instrument = instrument;
+            this.specialization = specialization;
+            this.derivation = derivation;
+            this.entity = entity;
+            this.association = association;
+            this.generation = generation;
+            this.activity = activity;
         }
     }