+++ modules-core/prov-model/src/main/java/org/openprovenance/prov/model/HasLocation.java
@@ -35,7 +35,6 @@ public interface HasLocation {
      * 
      * 
      */
-
-    public List<Location> getLocation();
+    List<Location> getLocation();
 
 }