+++ RELEASES.md
@@ -51,6 +51,7 @@ See also GitHub releases https://github.com/lucmoreau/ProvToolbox/releases
 * prov-jsonld: fixed incompatibility issues with jsonld1.1
 * prov-service-security-pac: created sub-module
 * prov-dot: displays qualification patterns
+* prov-model: extension of builder to support book example
 
 `2.0.5`
 * prov-template-library: cleaned up pom.xml